# in place of #39: change function call; optimize Lorenzo kernels