--- a/src/cusz_interface.cu
+++ b/src/cusz_interface.cu
@@ -2,9 +2,9 @@
  * @file cusz_workflow.cu
  * @author Jiannan Tian
  * @brief Workflow of cuSZ.
- * @version 0.1
- * @date 2020-09-20
- * Created on: 2020-02-12
+ * @version 0.2
+ * @date 2021-01-16
+ * (create) 2020-02-12; (release) 2020-09-20; (rev1) 2021-01-16
  *
  * @copyright (C) 2020 by Washington State University, The University of Alabama, Argonne National Laboratory
  * See LICENSE in top-level directory
@@ -57,32 +57,56 @@ void cusz::impl::PdQ(Data* d_d, Quant* d_q, size_t* dims, double* eb_variants)
     auto  d_eb_variants = mem::CreateDeviceSpaceAndMemcpyFromHost(eb_variants, 4);
     void* args[]        = {&d_d, &d_q, &d_dims, &d_eb_variants};
 
+    lorenzo_zip ctx;
+    void*       args_v2[] = {&ctx, &d_d, &d_q};
+    {
+        ctx.d0     = dims[DIM0];
+        ctx.d1     = dims[DIM1];
+        ctx.d2     = dims[DIM2];
+        ctx.radius = dims[RADIUS];
+        ctx.ebx2_r = eb_variants[EBx2_r];
+
+        ctx.stride1 = ctx.d0;
+        ctx.stride2 = ctx.d0 * ctx.d1;
+    }
+
     if (dims[nDIM] == 1) {
         static const int B = MetadataTrait<1>::Block;
 
-        dim3 block_num(dims[nBLK0]);
-        dim3 thread_num(B);
+        dim3 block_num(dims[nBLK0]), thread_num(B);
+
+        // cudaLaunchKernel(
+        //    (void*)cusz::predictor_quantizer::c_lorenzo_1d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
+
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::c_lorenzo_1d1l<Data, Quant>,  //
-            block_num, thread_num, args, 0, nullptr);
+            (void*)cusz::predictor_quantizer::c_lorenzo_1d1l_v2<Data, Quant>,  //
+            block_num, thread_num, args_v2, 0, nullptr);
     }
     else if (dims[nDIM] == 2) {
         static const int B = MetadataTrait<2>::Block;
 
-        dim3 block_num(dims[nBLK0], dims[nBLK1]);
-        dim3 thread_num(B, B);
+        dim3 block_num(dims[nBLK0], dims[nBLK1]), thread_num(B, B);
+
+        // cudaLaunchKernel(
+        //    (void*)cusz::predictor_quantizer::c_lorenzo_2d1l<Data, Quant>, block_num, thread_num, args,
+        //    (B + 1) * (B + 1) * sizeof(Data), nullptr);
+
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::c_lorenzo_2d1l<Data, Quant>,  //
-            block_num, thread_num, args, (B + 1) * (B + 1) * sizeof(Data), nullptr);
+            (void*)cusz::predictor_quantizer::c_lorenzo_2d1l_v2<Data, Quant>,  //
+            block_num, thread_num, args_v2, (B + 1) * (B + 1) * sizeof(Data), nullptr);
     }
     else if (dims[nDIM] == 3) {
         static const int B = MetadataTrait<3>::Block;
 
-        dim3 block_num(dims[nBLK0], dims[nBLK1], dims[nBLK2]);
-        dim3 thread_num(B, B, B);
+        dim3 block_num(dims[nBLK0], dims[nBLK1], dims[nBLK2]), thread_num(B, B, B);
+
+        // cudaLaunchKernel(
+        //    (void*)cusz::predictor_quantizer::c_lorenzo_3d1l<Data, Quant>, block_num, thread_num, args,
+        //    (B + 1) * (B + 1) * (B + 1) * sizeof(Data), nullptr);
+
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::c_lorenzo_3d1l<Data, Quant>,  //
-            block_num, thread_num, args, (B + 1) * (B + 1) * (B + 1) * sizeof(Data), nullptr);
+            (void*)cusz::predictor_quantizer::c_lorenzo_3d1l_v2<Data, Quant>,  //
+            block_num, thread_num, args_v2, (B + 1) * (B + 1) * (B + 1) * sizeof(Data), nullptr);
     }
     HANDLE_ERROR(cudaDeviceSynchronize());
 }
@@ -93,13 +117,33 @@ void cusz::impl::ReversedPdQ(Data* d_xd, Quant* d_q, Data* d_outlier, size_t* di
     auto  d_dims = mem::CreateDeviceSpaceAndMemcpyFromHost(dims, 16);
     void* args[] = {&d_xd, &d_outlier, &d_q, &d_dims, &_2eb};
 
+    lorenzo_unzip ctx;
+    {
+        ctx.d0      = dims[DIM0];
+        ctx.d1      = dims[DIM1];
+        ctx.d2      = dims[DIM2];
+        ctx.n_blk0  = dims[nBLK0];
+        ctx.n_blk1  = dims[nBLK1];
+        ctx.n_blk2  = dims[nBLK2];
+        ctx.radius  = dims[RADIUS];
+        ctx.ebx2    = _2eb;
+        ctx.stride1 = ctx.d0;
+        ctx.stride2 = ctx.d0 * ctx.d1;
+    }
+    void* args_v2[] = {&ctx, &d_xd, &d_outlier, &d_q};
+
     if (dims[nDIM] == 1) {
         static const int p = MetadataTrait<1>::Block;
 
         dim3 thread_num(p);
         dim3 block_num((dims[nBLK0] - 1) / p + 1);
+        // cudaLaunchKernel(
+        // (void*)cusz::predictor_quantizer::x_lorenzo_1d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
+
+        // v2
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::x_lorenzo_1d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
+            (void*)cusz::predictor_quantizer::x_lorenzo_1d1l_v2<Data, Quant>, block_num, thread_num, args_v2, 0,
+            nullptr);
     }
     else if (dims[nDIM] == 2) {
         const static size_t p = MetadataTrait<2>::Block;
@@ -108,8 +152,12 @@ void cusz::impl::ReversedPdQ(Data* d_xd, Quant* d_q, Data* d_outlier, size_t* di
         dim3 block_num(
             (dims[nBLK0] - 1) / p + 1,   //
             (dims[nBLK1] - 1) / p + 1);  //
+
+        // cudaLaunchKernel(
+        // (void*)cusz::predictor_quantizer::x_lorenzo_2d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::x_lorenzo_2d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
+            (void*)cusz::predictor_quantizer::x_lorenzo_2d1l_v2<Data, Quant>, block_num, thread_num, args_v2, 0,
+            nullptr);
     }
     else if (dims[nDIM] == 3) {
         const static size_t p = MetadataTrait<3>::Block;
@@ -119,8 +167,12 @@ void cusz::impl::ReversedPdQ(Data* d_xd, Quant* d_q, Data* d_outlier, size_t* di
             (dims[nBLK0] - 1) / p + 1,   //
             (dims[nBLK1] - 1) / p + 1,   //
             (dims[nBLK2] - 1) / p + 1);  //
+
+        // cudaLaunchKernel(
+        // (void*)cusz::predictor_quantizer::x_lorenzo_3d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::x_lorenzo_3d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
+            (void*)cusz::predictor_quantizer::x_lorenzo_3d1l_v2<Data, Quant>, block_num, thread_num, args_v2, 0,
+            nullptr);
     }
     else {
         cerr << log_err << "no 4D" << endl;