--- a/src/cusz_interface.cu
+++ b/src/cusz_interface.cu
@@ -47,75 +47,40 @@ using std::cout;
 using std::endl;
 using std::string;
 
-// typedef std::tuple<size_t, size_t, size_t, bool> tuple3ul;
-typedef std::tuple<size_t, size_t, size_t, bool> tuple_3ul_1bool;
-
 template <typename Data, typename Quant>
 void cusz::impl::PdQ(Data* d_d, Quant* d_q, size_t* dims, double* eb_variants)
 {
-    auto d_dims        = mem::CreateDeviceSpaceAndMemcpyFromHost(dims, 16);
-    auto d_eb_variants = mem::CreateDeviceSpaceAndMemcpyFromHost(eb_variants, 4);
-    // void* args[]        = {&d_d, &d_q, &d_dims, &d_eb_variants};
-
     lorenzo_zip ctx;
-    void*       args_v2v3[] = {&ctx, &d_d, &d_q};
+    void*       lorenzo_args[] = {&ctx, &d_d, &d_q};
     {
-        ctx.d0     = dims[DIM0];
-        ctx.d1     = dims[DIM1];
-        ctx.d2     = dims[DIM2];
-        ctx.radius = dims[RADIUS];
-        ctx.ebx2_r = eb_variants[EBx2_r];
-
-        ctx.stride1 = ctx.d0;
-        ctx.stride2 = ctx.d0 * ctx.d1;
+        ctx.d0 = dims[DIM0], ctx.d1 = dims[DIM1], ctx.d2 = dims[DIM2];
+        ctx.radius = dims[RADIUS], ctx.ebx2_r = eb_variants[EBx2_r];
+        ctx.stride1 = ctx.d0, ctx.stride2 = ctx.d0 * ctx.d1;
     }
 
     if (dims[nDIM] == 1) {
         static const int B = MetadataTrait<1>::Block;
 
         dim3 block_num(dims[nBLK0]), thread_num(B);
-
-        // cudaLaunchKernel(
-        //    (void*)cusz::predictor_quantizer::c_lorenzo_1d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
-
-        // for 1D, simply use v2
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::c_lorenzo_1d1l_v2<Data, Quant>,  //
-            block_num, thread_num, args_v2v3, 0, nullptr);
+            (void*)cusz::predictor_quantizer::v2::c_lorenzo_1d1l<Data, Quant>,  //
+            block_num, thread_num, lorenzo_args, 0, nullptr);
     }
     else if (dims[nDIM] == 2) {
         static const int B = MetadataTrait<2>::Block;
 
         dim3 block_num(dims[nBLK0], dims[nBLK1]), thread_num(B, B);
-
-        // cudaLaunchKernel(
-        //    (void*)cusz::predictor_quantizer::c_lorenzo_2d1l<Data, Quant>, block_num, thread_num, args,
-        //    (B + 1) * (B + 1) * sizeof(Data), nullptr);
-
-        // cudaLaunchKernel(
-        //     (void*)cusz::predictor_quantizer::c_lorenzo_2d1l_v2<Data, Quant>,  //
-        //     block_num, thread_num, args_v2v3, (B + 1) * (B + 1) * sizeof(Data), nullptr);
-
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::c_lorenzo_2d1l_v3<Data, Quant>,  //
-            block_num, thread_num, args_v2v3, B * B * sizeof(Data), nullptr);
+            (void*)cusz::predictor_quantizer::v3::c_lorenzo_2d1l<Data, Quant>,  //
+            block_num, thread_num, lorenzo_args, B * B * sizeof(Data), nullptr);
     }
     else if (dims[nDIM] == 3) {
         static const int B = MetadataTrait<3>::Block;
 
         dim3 block_num(dims[nBLK0], dims[nBLK1], dims[nBLK2]), thread_num(B, B, B);
-
-        // cudaLaunchKernel(
-        //    (void*)cusz::predictor_quantizer::c_lorenzo_3d1l<Data, Quant>, block_num, thread_num, args,
-        //    (B + 1) * (B + 1) * (B + 1) * sizeof(Data), nullptr);
-
-        // cudaLaunchKernel(
-        //     (void*)cusz::predictor_quantizer::c_lorenzo_3d1l_v2<Data, Quant>,  //
-        //     block_num, thread_num, args_v2v3, (B + 1) * (B + 1) * (B + 1) * sizeof(Data), nullptr);
-
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::c_lorenzo_3d1l_v3<Data, Quant>,  //
-            block_num, thread_num, args_v2v3, B * B * B * sizeof(Data), nullptr);
+            (void*)cusz::predictor_quantizer::v3::c_lorenzo_3d1l<Data, Quant>,  //
+            block_num, thread_num, lorenzo_args, B * B * B * sizeof(Data), nullptr);
     }
     HANDLE_ERROR(cudaDeviceSynchronize());
 }
@@ -123,72 +88,44 @@ void cusz::impl::PdQ(Data* d_d, Quant* d_q, size_t* dims, double* eb_variants)
 template <typename Data, typename Quant>
 void cusz::impl::ReversedPdQ(Data* d_xd, Quant* d_q, Data* d_outlier, size_t* dims, double _2eb)
 {
-    auto  d_dims = mem::CreateDeviceSpaceAndMemcpyFromHost(dims, 16);
-    // void* args[] = {&d_xd, &d_outlier, &d_q, &d_dims, &_2eb};
-
     lorenzo_unzip ctx;
     {
-        ctx.d0      = dims[DIM0];
-        ctx.d1      = dims[DIM1];
-        ctx.d2      = dims[DIM2];
-        ctx.n_blk0  = dims[nBLK0];
-        ctx.n_blk1  = dims[nBLK1];
-        ctx.n_blk2  = dims[nBLK2];
-        ctx.radius  = dims[RADIUS];
-        ctx.ebx2    = _2eb;
-        ctx.stride1 = ctx.d0;
-        ctx.stride2 = ctx.d0 * ctx.d1;
+        ctx.d0 = dims[DIM0], ctx.d1 = dims[DIM1], ctx.d2 = dims[DIM2];
+        ctx.n_blk0 = dims[nBLK0], ctx.n_blk1 = dims[nBLK1], ctx.n_blk2 = dims[nBLK2];
+        ctx.radius = dims[RADIUS], ctx.ebx2 = _2eb;
+        ctx.stride1 = ctx.d0, ctx.stride2 = ctx.d0 * ctx.d1;
     }
-    void* args_v2v3[] = {&ctx, &d_xd, &d_outlier, &d_q};
+    void* lorenzo_args[] = {&ctx, &d_xd, &d_outlier, &d_q};
 
     if (dims[nDIM] == 1) {
         static const int p = MetadataTrait<1>::Block;
 
-        dim3 thread_num(p);
-        dim3 block_num((dims[nBLK0] - 1) / p + 1);
-        // cudaLaunchKernel(
-        // (void*)cusz::predictor_quantizer::x_lorenzo_1d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
-
-        // v2
+        dim3 thread_num(p), block_num((dims[nBLK0] - 1) / p + 1);
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::x_lorenzo_1d1l_v2<Data, Quant>, block_num, thread_num, args_v2v3, 0,
-            nullptr);
+            (void*)cusz::predictor_quantizer::v2::x_lorenzo_1d1l<Data, Quant>,  //
+            block_num, thread_num, lorenzo_args, 0, nullptr);
     }
     else if (dims[nDIM] == 2) {
         const static size_t p = MetadataTrait<2>::Block;
 
-        dim3 thread_num(p, p);
-        dim3 block_num(
-            (dims[nBLK0] - 1) / p + 1,   //
-            (dims[nBLK1] - 1) / p + 1);  //
-
-        // cudaLaunchKernel(
-        // (void*)cusz::predictor_quantizer::x_lorenzo_2d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
+        dim3 thread_num(p, p), block_num((dims[nBLK0] - 1) / p + 1, (dims[nBLK1] - 1) / p + 1);
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::x_lorenzo_2d1l_v2<Data, Quant>, block_num, thread_num, args_v2v3, 0,
-            nullptr);
+            (void*)cusz::predictor_quantizer::v2::x_lorenzo_2d1l<Data, Quant>,  //
+            block_num, thread_num, lorenzo_args, 0, nullptr);
     }
     else if (dims[nDIM] == 3) {
         const static size_t p = MetadataTrait<3>::Block;
 
         dim3 thread_num(p, p, p);
-        dim3 block_num(
-            (dims[nBLK0] - 1) / p + 1,   //
-            (dims[nBLK1] - 1) / p + 1,   //
-            (dims[nBLK2] - 1) / p + 1);  //
-
-        // cudaLaunchKernel(
-        // (void*)cusz::predictor_quantizer::x_lorenzo_3d1l<Data, Quant>, block_num, thread_num, args, 0, nullptr);
+        dim3 block_num((dims[nBLK0] - 1) / p + 1, (dims[nBLK1] - 1) / p + 1, (dims[nBLK2] - 1) / p + 1);
         cudaLaunchKernel(
-            (void*)cusz::predictor_quantizer::x_lorenzo_3d1l_v2<Data, Quant>, block_num, thread_num, args_v2v3, 0,
-            nullptr);
+            (void*)cusz::predictor_quantizer::v2::x_lorenzo_3d1l<Data, Quant>,  //
+            block_num, thread_num, lorenzo_args, 0, nullptr);
     }
     else {
         cerr << log_err << "no 4D" << endl;
     }
     cudaDeviceSynchronize();
-
-    cudaFree(d_dims);
 }
 
 template <typename Data, typename Quant>
@@ -200,8 +137,7 @@ void cusz::impl::VerifyHuffman(
     size_t*       dims,
     double*       eb_variants)
 {
-    // TODO error handling from invalid read
-    cout << log_info << "Redo PdQ just to get quantization dump." << endl;
+    LogAll(log_info, "Redo PdQ just to get quant data.");
 
     auto  veri_data   = io::ReadBinaryToNewArray<Data>(fi, len);
     Data* veri_d_data = mem::CreateDeviceSpaceAndMemcpyFromHost(veri_data, len);