@dingwentao 2021-01-17
> HACC	GB/s			CESM	GB/s			NYX	GB/s	
> v1		3.60 ms	290.7	100.0%		96.83 us	249.3	100.0%		2.48 ms	201.6	100.0%
> v2		3.58 ms	292.4	100.6%		92.83 us	260.0	104.3%		2.27 ms	220.3	109.3%
> v3		-	-	-		88.83 us	271.8	109.0%		2.02 ms	247.5	122.8%

You also mentioned decompression speed is improved. I suggest adding decompression throughputs at some point.