@dingwentao 2021-01-18
> @dingwentao The decompression throughput is inconsistent across GPUs. Need to investigate.
> Let me keep `optim` for now, there will be some minor tunings.

Ok, sounds good. Thanks.